realms work in progress
# Conflicts:
#	Source/ACE.Entity/Enum/Properties/PropertyInt.cs

(cherry picked from commit 2844768074666f50171e929335f4d30d0d2597b2)

+++ Source/ACE.Server/Physics/Common/LScape.cs
@@ -47,7 +47,7 @@ namespace ACE.Server.Physics.Common
 
         public static int LandblocksCount => Landblocks.Count;
 
-        public static Landblock get_landblock(uint blockCellID, byte instance)
+        public static Landblock get_landblock(uint blockCellID, uint instance)
         {
             var iBlockCell = BlockCell.GetLongCell(blockCellID, instance);
